DjangoAPI/requirements.txt
django==3.0.2
psycopg2==2.8.4
djangorestframework==3.11.0